refactor: separate secrets module
